+++ src/main/java/com/jayway/security/SecureAccountServiceImpl.java
@@ -25,49 +25,49 @@ class SecureAccountServiceImpl implements AccountService {
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public ImmutableAccount get(Long accountNumber) throws UnknownAccountException {
+    public ImmutableAccount get(Integer accountNumber) throws UnknownAccountException {
         return accountService.get(accountNumber);
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public void deposit(Long accountNumber, long amount) throws UnknownAccountException {
+    public void deposit(Integer accountNumber, int amount) throws UnknownAccountException {
         accountService.deposit(accountNumber, amount);
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public ImmutableAccount withdraw(Long accountNumber, long amount) throws UnknownAccountException {
+    public ImmutableAccount withdraw(Integer accountNumber, int amount) throws UnknownAccountException {
         return accountService.withdraw(accountNumber, amount);
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public void transfer(Long fromAccountNumber, Long toAccountNumber, long amount) throws UnknownAccountException {
+    public void transfer(Integer fromAccountNumber, Integer toAccountNumber, int amount) throws UnknownAccountException {
         accountService.transfer(fromAccountNumber, toAccountNumber, amount);
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public Long createAccount() {
+    public Integer createAccount() {
         return accountService.createAccount();
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public void deleteAccount(Long accountNumber) throws UnknownAccountException {
+    public void deleteAccount(Integer accountNumber) throws UnknownAccountException {
         accountService.deleteAccount(accountNumber);
     }
 
 
     @PreAuthorize("hasAuthority('ACCOUNT_OWNER')")
     @Override
-    public List<Long> getAllAccountNumbers() {
+    public List<Integer> getAllAccountNumbers() {
         return accountService.getAllAccountNumbers();
     }
 }
