+++ src/main/resources/db-schema.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS account_t;
 
 CREATE TABLE account_t (
-  account_number BIGINT NOT NULL AUTO_INCREMENT,
-  balance BIGINT NOT NULL,
+  account_number INT NOT NULL AUTO_INCREMENT,
+  balance INT NOT NULL,
   PRIMARY KEY (account_number),
   CHECK (balance >= 0)
 );
